Extract HOLD_STATE_MAP to shared utility, lazy load emoji picker, improve error handling
- Deduplicate HOLD_STATE_MAP and convertLitUpHoldsStringToMap into
  packages/backend/src/db/queries/util/hold-state.ts, fixing incorrect
  MoonBoard hold codes (1,2,3 -> 42,43,44) in favorites and playlists
- Lazy load emoji-mart picker via next/dynamic to reduce initial bundle
  size (~500KB deferred until user opens emoji popover)
- Add snackbar error notifications in liked-climbs-list and preserve
  already-loaded data when subsequent page fetches fail
