+++ packages/backend/src/db/queries/climbs/get-climb.ts
@@ -1,84 +1,8 @@
 import { sql } from 'drizzle-orm';
 import { db } from '../../client';
 import { getBoardTables, type BoardName } from '../util/table-select';
-import type { Climb, LitUpHoldsMap, HoldState } from '@boardsesh/shared-schema';
-
-// Hold state mapping for converting frames string to lit up holds map
-type HoldColor = string;
-type HoldCode = number;
-
-const HOLD_STATE_MAP: Record<
-  BoardName,
-  Record<HoldCode, { name: HoldState; color: HoldColor; displayColor?: HoldColor }>
-> = {
-  kilter: {
-    42: { name: 'STARTING', color: '#00FF00' },
-    43: { name: 'HAND', color: '#00FFFF' },
-    44: { name: 'FINISH', color: '#FF00FF' },
-    45: { name: 'FOOT', color: '#FFAA00' },
-    12: { name: 'STARTING', color: '#00FF00' },
-    13: { name: 'HAND', color: '#00FFFF' },
-    14: { name: 'FINISH', color: '#FF00FF' },
-    15: { name: 'FOOT', color: '#FFAA00' },
-  },
-  tension: {
-    1: { name: 'STARTING', displayColor: '#00DD00', color: '#00FF00' },
-    2: { name: 'HAND', displayColor: '#4444FF', color: '#0000FF' },
-    3: { name: 'FINISH', displayColor: '#FF0000', color: '#FF0000' },
-    4: { name: 'FOOT', displayColor: '#FF00FF', color: '#FF00FF' },
-    5: { name: 'STARTING', displayColor: '#00DD00', color: '#00FF00' },
-    6: { name: 'HAND', displayColor: '#4444FF', color: '#0000FF' },
-    7: { name: 'FINISH', displayColor: '#FF0000', color: '#FF0000' },
-    8: { name: 'FOOT', displayColor: '#FF00FF', color: '#FF00FF' },
-  },
-  moonboard: {
-    42: { name: 'STARTING', color: '#00FF00', displayColor: '#44FF44' },
-    43: { name: 'HAND', color: '#0000FF', displayColor: '#4444FF' },
-    44: { name: 'FINISH', color: '#FF0000', displayColor: '#FF3333' },
-  },
-};
-
-// Warned hold states to avoid log spam
-const warnedHoldStates = new Set<string>();
-
-/**
- * Convert lit up holds string to a map
- * Returns only the first frame for single-frame climbs
- */
-function convertLitUpHoldsStringToMap(litUpHolds: string, board: BoardName): Record<number, LitUpHoldsMap> {
-  return litUpHolds
-    .split(',')
-    .filter((frame) => frame)
-    .reduce(
-      (frameMap, frameString, frameIndex) => {
-        const frameHoldsMap = Object.fromEntries(
-          frameString
-            .split('p')
-            .filter((hold) => hold)
-            .map((holdData) => holdData.split('r').map((str) => Number(str)))
-            .map(([holdId, stateCode]) => {
-              const stateInfo = HOLD_STATE_MAP[board]?.[stateCode];
-              if (!stateInfo) {
-                // Rate-limit warnings to avoid log spam
-                const warnKey = `${board}:${stateCode}`;
-                if (!warnedHoldStates.has(warnKey)) {
-                  warnedHoldStates.add(warnKey);
-                  console.warn(
-                    `HOLD_STATE_MAP is missing values for ${board} status code: ${stateCode} (this warning is only shown once per status code)`,
-                  );
-                }
-                return [holdId || 0, { state: `${holdId}=${stateCode}` as HoldState, color: '#FFF', displayColor: '#FFF' }];
-              }
-              const { name, color, displayColor } = stateInfo;
-              return [holdId, { state: name, color, displayColor: displayColor || color }];
-            }),
-        );
-        frameMap[frameIndex] = frameHoldsMap as LitUpHoldsMap;
-        return frameMap;
-      },
-      {} as Record<number, LitUpHoldsMap>,
-    );
-}
+import type { Climb } from '@boardsesh/shared-schema';
+import { convertLitUpHoldsStringToMap } from '../util/hold-state';
 
 interface GetClimbParams {
   board_name: BoardName;
