+++ packages/backend/src/graphql/resolvers/playlists/queries.ts
@@ -13,71 +13,7 @@ import {
 } from '../../../validation/schemas';
 import { getBoardTables, isValidBoardName } from '../../../db/queries/util/table-select';
 import { getSizeEdges } from '../../../db/queries/util/product-sizes-data';
-import type { LitUpHoldsMap, HoldState } from '@boardsesh/shared-schema';
-
-// Hold state mapping for converting frames string to lit up holds map
-type HoldColor = string;
-type HoldCode = number;
-// BoardName is imported from @boardsesh/shared-schema
-
-const HOLD_STATE_MAP: Record<
-  BoardName,
-  Record<HoldCode, { name: HoldState; color: HoldColor; displayColor?: HoldColor }>
-> = {
-  kilter: {
-    42: { name: 'STARTING', color: '#00FF00' },
-    43: { name: 'HAND', color: '#00FFFF' },
-    44: { name: 'FINISH', color: '#FF00FF' },
-    45: { name: 'FOOT', color: '#FFAA00' },
-    12: { name: 'STARTING', color: '#00FF00' },
-    13: { name: 'HAND', color: '#00FFFF' },
-    14: { name: 'FINISH', color: '#FF00FF' },
-    15: { name: 'FOOT', color: '#FFAA00' },
-  },
-  tension: {
-    1: { name: 'STARTING', displayColor: '#00DD00', color: '#00FF00' },
-    2: { name: 'HAND', displayColor: '#4444FF', color: '#0000FF' },
-    3: { name: 'FINISH', displayColor: '#FF0000', color: '#FF0000' },
-    4: { name: 'FOOT', displayColor: '#FF00FF', color: '#FF00FF' },
-    5: { name: 'STARTING', displayColor: '#00DD00', color: '#00FF00' },
-    6: { name: 'HAND', displayColor: '#4444FF', color: '#0000FF' },
-    7: { name: 'FINISH', displayColor: '#FF0000', color: '#FF0000' },
-    8: { name: 'FOOT', displayColor: '#FF00FF', color: '#FF00FF' },
-  },
-  // MoonBoard uses a different hold format - holds are identified by grid position
-  moonboard: {
-    1: { name: 'STARTING', color: '#00FF00' },
-    2: { name: 'HAND', color: '#0000FF' },
-    3: { name: 'FINISH', color: '#FF0000' },
-  },
-};
-
-function convertLitUpHoldsStringToMap(litUpHolds: string, board: BoardName): Record<number, LitUpHoldsMap> {
-  return litUpHolds
-    .split(',')
-    .filter((frame) => frame)
-    .reduce(
-      (frameMap, frameString, frameIndex) => {
-        const frameHoldsMap = Object.fromEntries(
-          frameString
-            .split('p')
-            .filter((hold) => hold)
-            .map((holdData) => holdData.split('r').map((str) => Number(str)))
-            .map(([holdId, stateCode]) => {
-              const stateInfo = HOLD_STATE_MAP[board]?.[stateCode];
-              if (!stateInfo) {
-                return [holdId || 0, { state: `${holdId}=${stateCode}` as HoldState, color: '#FFF', displayColor: '#FFF' }];
-              }
-              const { name, color, displayColor } = stateInfo;
-              return [holdId, { state: name, color, displayColor: displayColor || color }];
-            }),
-        );
-        frameMap[frameIndex] = frameHoldsMap as LitUpHoldsMap;
-        return frameMap;
-      },
-      {} as Record<number, LitUpHoldsMap>,
-    );
-}
+import { convertLitUpHoldsStringToMap } from '../../../db/queries/util/hold-state';
 
 export const playlistQueries = {
   /**
